attempt to import all actions
