+++ src/mlde_utils/data/constrain.py
@@ -1,6 +1,6 @@
 import logging
 
-from mlde_utils.data import register_action
+from mlde_utils.data.actions_registry import register_action
 
 logger = logging.getLogger(__name__)
