+++ src/mlde_utils/data/resample.py
@@ -1,7 +1,7 @@
 import datetime
 import logging
 
-from mlde_utils.data import register_action
+from mlde_utils.data.actions_registry import register_action
 
 logger = logging.getLogger(__name__)
