+++ src/mlde_utils/data/vorticity.py
@@ -2,7 +2,7 @@ import logging
 
 import metpy.calc as mpcalc
 
-from mlde_utils.data import register_action
+from mlde_utils.data.actions_registry import register_action
 
 logger = logging.getLogger(__name__)
